udpated view portal ts file

+++ src/app/view-portal/view-portal.component.ts
@@ -3,19 +3,37 @@ import { ViewPortalService, Portal } from "../service/view-portal.service";
 import { DialogBoxService } from "../service/dialog-box.service";
 import { DialogBoxComponent} from "../dialog-box/dialog-box.component";
 import { Router } from '@angular/router';
+import { FormGroup } from '@angular/forms';
+import {  FormControl, FormBuilder, Validators } from '@angular/forms';
 import {MatTableDataSource, MatSort,MatDialog ,MatPaginator} from '@angular/material';
 
-
+export class ImageLocation {
+  constructor(
+    public pathId: number,
+    public portalID: number,
+    public imgPath: string,
+    public folderName: string,
+    public imageName: string
+    
+  ) {}
+}
 @Component({
   selector: "app-portals",
-  templateUrl: "./view-portal.component.html",
-  styleUrls: ["./view-portal.component.css"]
+  templateUrl: "./view-portal.component.html"
 })
+
 export class ViewPortalComponent implements OnInit {
+  form: FormGroup = new FormGroup({
+    browser: new FormControl('')})
   portal: Portal[];
-  
+  imgList:any;
+  imageLocation: ImageLocation[];
+  browservalue="Chrome"
+  Browser: any = ['FireFox']
+  hide=true;
   isPopupOpened = true;
-  displayedColumns: string[] = ['PortalID','Industry','ClientName','Navigation', 'PortalName', 'PortalUrl','ModuleName', 'Fields','SampleData','Delete','Update','Run'];
+  Imagepath: any=[]
+  displayedColumns: string[] = ['PortalID','BusinessUnit','SubBusinessUnit','ClientName','PortalName', 'PortalUrl','ModuleName', 'Fields','Status','Result','Run'];
   portalInfo = new MatTableDataSource(this.portal);
 
   constructor(private service: ViewPortalService,private dialog: MatDialog,private router: Router,
@@ -32,6 +50,11 @@ export class ViewPortalComponent implements OnInit {
   });
   this.portalInfo.paginator = this.paginator;
     this.portalInfo.sort = this.sort;
+  }
+  changeDrop(e,drop){
+    var dropvalue=e.target.value;
+    this.browservalue=dropvalue;
+    console.log(this.browservalue)
   }
   editPortal(id: number) {
     this.isPopupOpened = true;
@@ -67,8 +90,22 @@ export class ViewPortalComponent implements OnInit {
       console.log("failure"+err.log);
     });
 }
+
 deletPortal(portal): void {
+  if(confirm("Are you sure to delete ")) {
     this.service.deletePortal(portal).subscribe((data)=>{
+      if(data){
+        window.location.reload()
+        this.router.navigate(["/viewPortal"])
+      }
+    },(err)=>{
+      console.log("failure"+err.log);
+    });
+  }
+  }
+  navigatePortal(portal): void {
+
+    this.service.navigatePortal(portal,this.browservalue).subscribe((data)=>{
       if(data){
         window.location.reload()
         this.router.navigate(["/viewPortal"])
@@ -86,5 +123,3 @@ export interface PeriodicElement {
   fields: string;
   action: string;
 }
-
-
